StructureAnimation is almost working correctly, but subsequent renders aren't calling updateTree
