+++ components/GitAnimation.js
@@ -4,7 +4,7 @@ export default class GitAnimation extends Component {
   render() {
     return (
       //JSX
-      <div>Git Animation</div>
+      <div>We will show a Git animation here</div>
     )
   }
 }
